requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.0
cloudinary==1.28.0
confusable-homoglyphs==3.2.0
Django==3.2.9
django-bootstrap-v5==1.0.7
django-bootstrap5==21.3
django-registration==3.1.2
django-registration-redux==2.10
djangorestframework==3.13.1
importlib-metadata==3.10.1
Pillow==9.0.1
python-decouple==3.6
pytz==2022.1
soupsieve==2.3.2
sqlparse==0.4.2
zipp==3.8.0
whitenoise==5.3.0
psycopg2==2.9.3
dj-database-url==0.5.0
requests==2.26.0